hjepa/requirements.txt
alabaster==0.7.13
alembic==1.4.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asciitree==0.3.3
asttokens==2.2.1
async-lru==2.0.4
attrs==22.2.0
azure-core==1.26.0
azure-storage-blob==12.14.1
Babel==2.12.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.1
black==23.3.0
bleach==6.0.0
blessings==1.7
bokeh
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==1.2.2
cmake==3.26.1
comm==0.1.3
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
D4RL @ git+https://github.com/Farama-Foundation/d4rl@71a9549f2091accff93eeff68f1f3ab2c0e0a288
databricks-cli==0.17.3
debugpy==1.6.7
defusedxml==0.7.1
distlib==0.3.6
docutils==0.18.1
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
fancycompleter==0.9.1
Farama-Notifications==0.0.4
fastjsonschema==2.16.3
ffmpeg==1.4
filelock==3.12.0
flake8==6.0.0
flatten-dict==0.4.2
fonttools==4.39.3
fqdn==1.5.1
fsspec==2023.12.2
gdown==4.7.1
gpustat==0.6.0
gunicorn==20.1.0
html-testRunner==1.2.1
identify==2.5.22
idna==3.4
imagesize==1.4.1
importlib-metadata==4.13.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipdb==0.13.13
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isodate==0.6.1
isoduration==20.11.0
jax-jumpy==1.0.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.3
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.1
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab-vim==4.1.0
jupyterlab-widgets==3.0.7
jupyterlab_server==2.25.2
kiwisolver==1.4.4
kornia==0.6.5
lit==16.0.0
Mako==1.2.3
MarkupSafe==2.1.2
mccabe==0.7.0
metaworld @ git+https://github.com/rlworkgroup/metaworld.git@c822f28f582ba1ad49eb5dcf61016566f28003ba
mistune==2.0.5
mlflow==1.22.0
mpmath==1.3.0
msrest==0.7.1
mypy==1.2.0
mypy-extensions==1.0.0
natsort==8.4.0
nbclassic==0.5.5
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nbsphinx==0.9.1
nest-asyncio==1.5.6
networkx==3.1
nodeenv==1.7.0
notebook==6.5.3
notebook_shim==0.2.2
numcodecs==0.11.0
numpy==1.23.5
packaging==23.0
pandas==2.0.1
pandocfilters==1.5.0
paramiko==3.1.0
parso==0.8.3
patchelf==0.14.5.0
pathspec==0.11.1
pdbpp==0.10.3
pettingzoo==1.24.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pink-noise-rl==2.0.1
platformdirs==3.2.0
plotly==5.15.0
pluggy==1.0.0
pre-commit==3.2.2
prometheus-client==0.16.0
prometheus-flask-exporter==0.20.3
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pybullet==3.2.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pygame==2.3.0
pyglet==1.5.15
Pygments==2.14.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrepl==0.9.0
pyrsistent==0.19.3
pytest==7.2.2
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.7
pytorch-mppi==0.7.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
querystring-parser==1.2.4
referencing==0.32.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rlbench @ git+https://github.com/stepjam/RLBench.git@7c3f425f4a0b6b5ce001ba7246354eb3c70555be
rpds-py==0.15.2
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
Send2Trash==1.8.0
setproctitle==1.3.3
six==1.16.0
sk-video==1.1.10
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.4
Sphinx==6.1.3
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
stack-data==0.6.2
sympy==1.11.1
tenacity==8.2.2
tensorboard-plugin-wit==1.8.1
tinycss2==1.2.1
torch-tb-profiler==0.4.0
-e git+ssh://git@github.com/pytorch/rl.git@80c63ad40c7356559a2d4991215eb331ecff0f95#egg=torchrl
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transforms3d==0.4.1
triton==2.1.0
types-PyYAML==6.0.12.9
types-termcolor==1.1.6.2
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.18
virtualenv==20.22.0
wandb==0.15.12
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
wmctrl==0.5
xmltodict==0.13.0
xyzservices==2023.7.0
zarr==2.14.2
zipp==3.15.0
omegaconf==2.3.0
